cleanup custom signerToSafe7579 implementation
