--- advanced/wallets/react-wallet-v2/src/utils/safe7579AccountUtils/constants.ts
@@ -1,20 +1,0 @@
-import { ENTRYPOINT_ADDRESS_V07_TYPE } from 'permissionless/types'
-import { Address } from 'viem'
-
-export const ENTRY_POINT_ADDRESS: ENTRYPOINT_ADDRESS_V07_TYPE =
-  '0x0000000071727De22E5E9d8BAf0edAc6f37da032'
-
-export const SAFE_ACCOUNT_FACTORY_ADDRESS: Address = '0xe89e194e5bd3e5a8d40c4cd9c95dd2c56a8a6ed6'
-
-export const SAFE_SINGLETON_ADDRESS: Address = '0x8d70ae4ae3fb2a73e78f57da16e6e1edbe9fd3eb'
-
-export const SAFE_7579_ADDRESS: Address = '0xD2Aa2F19Bf6700bE99E140A0b283fE9EF9438A15'
-
-export const REGISTRY_ADDRESS: Address = '0x1d8c40f958fb6998067e9b8b26850d2ae30b7c70'
-
-export const LAUNCHPAD_ADDRESS: Address = '0x1EC6A1e000dD440995667e48cB880785C7d6831C'
-
-export const VALIDATOR_ADDRESS: Address = '0x503b54Ed1E62365F0c9e4caF1479623b08acbe77'
-
-export const SAFE7579_USER_OPERATION_BUILDER_ADDRESS: Address =
-  '0xBe9086ff28A0DEB285fa74fE9f290522AC2693EC'
